requirements.txt
batchgenerators==0.23
MedPy==0.4.0
monai==0.8.1
monai.egg==info
nibabel==3.2.2
numpy==1.22.2
pandas==1.3.5
PyYAML==6.0
ray==1.12.1
SimpleITK==2.1.1.2
torch==1.11.0a0+17540c5
tqdm==4.62.3
wandb==0.12.16
